% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drho_proposal.R
\name{drho_proposal}
\alias{drho_proposal}
\title{Density for the proposal distribution of \eqn{\rho}}
\usage{
drho_proposal(x, rho0)
}
\arguments{
\item{x}{a given proposed \eqn{\rho}.}

\item{rho0}{center point of the proposal density, i.e., having the highest
density, i.e. 1.}
}
\value{
a scale value
}
\description{
The density of \eqn{\rho} reaches its peak (with value 1) at \eqn{\rho_0}.
Then decreases (linearly) forward to 0 at point 1, and backward to 0 at
point -1.
}
\details{
The density for \eqn{\rho} is given as following
\deqn{f(x)=
\begin{cases}
\frac{x+1}{\rho_0+1}, &-1\le x\le\rho_0\\
\frac{1-x}{1-\rho_0}, &\rho_0<x\le 1
\end{cases}}
when \eqn{\rho_0\neq -1} and \eqn{\rho_0\neq 1}. For the two cases:
\itemize{
\item If \eqn{\rho_0=-1}, \eqn{f(x) = \frac{1-x}{2}, -1\le x\le 1};
\item If \eqn{\rho_0=1}, \eqn{f(x)=\frac{x+1}{2}, -1\le x\le 1}.
}

Putting together, for \eqn{\rho_0=-1} or 1,
\eqn{f(x) = \frac{1}{2} + \frac{\rho_0}{2}x, -1\le x \le 1.}
}
\examples{
drho_proposal(0,0.5)
drho_proposal(0,-0.5)
}
