% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_data}
\alias{gen_data}
\title{Generate and Write to Files Large Number of Sample Trajectories}
\usage{
gen_data(par_name, par, N, N_rep, h = 1, n_segment = 10)
}
\arguments{
\item{par_name}{parameter setting name, such as 'par0'.}

\item{par}{a vector of true values for the parameters
\eqn{\mu,k,\theta,\sigma_v,\rho}.}

\item{N}{the number of sample returns, \eqn{y_1,\cdots,y_N}.}

\item{N_rep}{the number of Sample Trajectories.}

\item{h}{time unit, defaults to 1.}

\item{n_segment}{number of further segments for each interval,
defaults to 10.}
}
\value{
no return.
}
\description{
Generate and Write to Files Large Number of Sample Trajectories. The files
are named according to 'par_name-N-ith.csv' where N denotes sample length.
}
\examples{
par0 = c(0.125,0.1,0.25,0.1,-0.7)
gen_data('par0', par0, N=100000, N_rep=2, h=1, n_segment=10)
}
