% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rho0.R
\name{get_rho0}
\alias{get_rho0}
\title{Get the sample correlation of the Brownian increments}
\usage{
get_rho0(v, y, mu, k, theta, sigma_v, h)
}
\arguments{
\item{v}{vector of volatility, [\eqn{v_0,v_1,\cdots,v_N}].}

\item{y}{vector of returns, [\eqn{y_1,\cdots,y_N}].}

\item{mu}{parameter \eqn{\mu}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}, the long-run mean of volatility.}

\item{sigma_v}{parameter \eqn{\sigma_v}, volatility of volatility.}

\item{h}{time unit.}
}
\value{
A scale value, \eqn{\rho_0}.
}
\description{
\eqn{\rho_0} is the sample correlation of the Brownian increments of
Heston Stochastic Volatility Model
\deqn{y_n = \mu h - \frac{1}{2}v_{n-1}h + \sqrt{v_{n-1}h}\epsilon_n^y}
\deqn{v_n - v_{n-1} = k(\theta - v_{n-1})h +\sigma_v\sqrt{v_{n-1}h}\epsilon_n^v}
\eqn{\epsilon_n^y} and \eqn{\epsilon_n^v} are first computed from return and
volatility samples.
\deqn{\epsilon_n^y = \frac{y_n-\mu h + \frac{1}{2}v_{n-1}h}{\sqrt{v_{n-1}h}}
,\qquad
\epsilon_n^v=\frac{v_n-v_{n-1}-k(\theta - v_{n-1})h}{\sigma_v\sqrt{v_{n-1}h}}.}
Then compute \eqn{\rho_0} as the sample correlation.
}
\examples{
v = rep(0.25,5)
y = rep(0.125,4)
mu = 0.125; k = 0.1; theta = 0.25; sigma_v = 0.1
h = 1
get_rho0(v,y,mu,k,theta,sigma_v,h)
}
