% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hestonmcmc-package.R
\docType{package}
\name{hestonmcmc-package}
\alias{hestonmcmc}
\alias{hestonmcmc-package}
\title{hestonmcmc: MCMC for Estimating Heston Stochastic Volatility Models}
\description{
An implementation of Markov Chain Monte Carlo (MCMC) method for estimating Heston Stochastic Volatility Models.
}
\details{
We try to implement the MCMC as close as to that in Eraker et al. (2003) and
Johannes and Polson (2010). However, there are some details are implemented
according to my judgment, because of the vague (maybe errors) in the two
papers.
}
\references{
Eraker, B., Johannes, M., & Polson, N. (2003). The impact of jumps in
volatility and returns. \emph{The Journal of Finance}, 58(3), 1269-1300.

Johannes, M., & Polson, N. (2010). MCMC methods for continuous-time
financial econometrics. In
\emph{Handbook of Financial Econometrics: Applications} (pp. 1-72). Elsevier.
}
\author{
Yanfeng Wu \href{mailto:yanfengwu13@fudan.edu.cn}{yanfengwu13@fudan.edu.cn}
}
\keyword{internal}
