% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mcmc.R
\name{cmcmc}
\alias{cmcmc}
\alias{mcmc}
\title{Estimate the parameters through MCMC}
\usage{
cmcmc(y, g = 5000L, G = 10000L, h = 1, echo = 0)

mcmc(y, g = 5000, G = 10000, h = 1, echo = FALSE)
}
\arguments{
\item{y}{vector of returns [\eqn{y_1, ..., y_N}].}

\item{g}{number of warm-up iteration, defaults to 5,000.}

\item{G}{number of total iteration, defaults to 10,000.}

\item{h}{time unit, defaults to 1.}

\item{echo}{TRUE or FALSE, whether or not echo the parameters and first 10
volatilities in the first and last ten iterations, defaults to FALSE.}
}
\value{
a list of two elements:
\itemize{
\item a vector of the estimated parameters, \eqn{(\mu,k,\theta,\sigma_v,\rho)},
\item a matrix record of volatility in which each row represents an
iteration result, i.e., with dimension of \eqn{G\times(N+1)}.
}
}
\description{
Given \eqn{[y_1,\cdots, y_N]}, to estimate parameters of Heston Stochastic
Volatility Model using MCMC.
\itemize{
\item \code{cmcmc()} implemented in C++, returns only a vector of the estimated
parameters.
\item \code{mcmc()} implemented in R.
}
}
\details{
Priors for and updating functions of these parameters:\tabular{ll}{
   prior \tab updating function \cr
   \eqn{\mu \sim \mathcal{N}(0,0.1^2) } \tab \code{\link[=rmu]{rmu()}} \cr
   \eqn{ k \sim \mathcal{N}(0.01,1^2) } \tab \code{\link[=rk]{rk()}} \cr
   \eqn{\theta \sim \mathcal{N}(0.01,1^2)} \tab \code{\link[=rtheta]{rtheta()}} \cr
   \eqn{\sigma_v \sim Gamma(shape=2,rate=1)} \tab \code{\link[=rsigma_v]{rsigma_v()}} \cr
   \eqn{\rho \sim Uniform(-1,1)} \tab \code{\link[=rrho]{rrho()}} \cr
}


Prior for the starting volatility: an uniform distribution over a large
range. (Depreciated
\eqn{v_0 \sim Gamma(shape=2k\theta/\sigma_v^2, rate=2k/\sigma_v^2)}.)

Updating functions of volatility:
\itemize{
\item \eqn{v_0}: \code{\link[=rv0]{rv0()}}
\item \eqn{v_n}: \code{\link[=rvi]{rvi()}} for \eqn{1\le n < N}
\item \eqn{v_N}: \code{\link[=rvN]{rvN()}}
}
}
\examples{
y = rep(0.125,20)
mcmc(y)
}
