% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rrho.R
\name{crrho}
\alias{crrho}
\alias{rrho}
\title{Update \eqn{\rho}}
\usage{
crrho(rho_old, v, y, mu, k, theta, sigma_v, h)

rrho(rho_old, v, y, mu, k, theta, sigma_v, h)
}
\arguments{
\item{rho_old}{value of \eqn{\rho} before updating, \eqn{\rho^{(g)}}, in
\eqn{g}th iteration.}

\item{v}{vector of volatility [\eqn{v_0, v_1, ..., v_N}].}

\item{y}{vector of returns [\eqn{y_1, ..., y_N}].}

\item{mu}{parameter \eqn{\mu}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma_v}{parameter \eqn{\sigma_v}.}

\item{h}{time unit.}
}
\value{
new value for \eqn{\rho}, a scale value.
}
\description{
Update \eqn{\rho}, using Independence Metropolis with proposal density
centered at the sample correlation between the Brownian increments.
\itemize{
\item \code{crrho()} implemented in C++, through package Rcpp.
\item \code{rrho()} implemented in R.
}
}
\section{Major steps}{

\enumerate{
\item First, compute the density center \eqn{\rho_0}
(\eqn{ = corr(\epsilon^y,\epsilon^v)}) of the proposal \eqn{P(\cdot)}, based
on samples [\eqn{v_0, v_1, ..., v_N}] and [\eqn{y_1, ..., y_N}] where
\deqn{\epsilon_n^y = \frac{y_n-\mu h+\frac{1}{2}v_{n-1}h}{\sqrt{v_{n-1}h}},
\qquad
\epsilon_n^v = \frac{v_n-v_{n-1}-k(\theta-v_{n-1})h}{\sigma_v\sqrt{v_{n-1}h}}.}
\item Then, propose a new \eqn{\rho^{(g+1)}} according to \eqn{P(\cdot)}, see
\code{\link[=rrho_proposal]{rrho_proposal()}}, \code{\link[=drho_proposal]{drho_proposal()}} for details.
\item Last, accept \eqn{\rho^{(g+1)}} with probability
\deqn{\alpha(\rho^{(g)},\rho^{(g+1)})
=min\left(\frac{\pi(\rho^{(g+1)})}{\pi(\rho^{(g)})}
         \times \frac{P(\rho^{(g)})}{P(\rho^{(g+1)})}, 1\right).}
}
}

\section{Posterior density ratio}{


Employ following version Heston SV model
\deqn{y_n=\mu h-\frac{1}{2}v_{n-1}h + \sqrt{v_{n-1}h}(\rho\epsilon_n^v
          +\sqrt{1-\rho^2}\epsilon_n)}
\deqn{v_n-v_{n-1}=k(\theta-v_{n-1})h + \sigma_v\sqrt{v_{n-1}h}\epsilon_n^v}
The posterior of \eqn{\rho}
\deqn{\begin{matrix}
      P(\rho|v_{0:N},y_{1:N})
      &\propto& P(y_{1:N}|v_{0:N})\cdot P(v_{0:N})\cdot P(\rho)\\
      &\propto& \prod_{n=1}^{N}P(y_n|v_{n-1},v_n)\cdot P(\rho)\\
      &\propto& \prod_{n=1}^{N}P(y_n|v_{n-1},v_n)\\
      &\propto& \prod_{n=1}^{N}P(x_n|v_{n-1},v_n)
\end{matrix}}
where \eqn{x_n\triangleq y_n-\mu h+\frac{1}{2}v_{n-1}h}, \eqn{P(\rho)} denote
the prior density, not the proposal density.
Note that \eqn{P(\rho)} is uniform(-1,1), thus constant.
\deqn{x_n|v_{n-1},v_n \sim \mathcal{N}(\rho\sqrt{v_{n-1}h}\epsilon_n^v,
      (1-\rho^2)v_{n-1}h)}
where
\deqn{\sqrt{v_{n-1}h}\epsilon_n^v =
      \frac{v_n-v_{n-1}-k(\theta-v_{n-1})h}{\sigma_v}.}
\deqn{P(x_n|v_{n-1},v_n)=\frac{1}{\sqrt{2\pi(1-\rho^2)v_{n-1}h}}
      e^{-\frac{(x_n-\rho\sqrt{v_{n-1}h}\epsilon_n^v)^2}{2(1-\rho^2)v_{n-1}h}}.}
Then
\deqn{\frac{\pi(\rho^{(g+1)})}{\pi(\rho^{(g)})} =
      \frac{P(\rho^{(g+1)}|v_{0:N},y_{1:N})}{P(\rho^{(g)}|v_{0:N},y_{1:N})}.}
\deqn{log\left(\frac{\pi(\rho^{(g+1)})}{\pi(\rho^{(g)})}\right)
      = - \frac{N}{2}log\frac{1-\rho^{(g+1)2}}{1-\rho^{(g)2}}
      - \sum_{n=1}^{N}\left[\frac{(x_n-\rho^{(g+1)}\sqrt{v_{n-1}h}\epsilon_n^v)^2}
      {2(1-\rho^{(g+1)2})v_{n-1}h}
      - \frac{(x_n-\rho^{(g)}\sqrt{v_{n-1}h}\epsilon_n^v)^2}
      {2(1-\rho^{(g)2})v_{n-1}h}\right]
      .}
Note that \eqn{\rho} can't be -1 or 1, otherwise there will produce \code{Inf}
and may produce \code{NaN}.
}

\examples{
rho_old = 0
v = rep(0.25,5)
y = rep(0.125,4)
mu = 0.125; k = 0.1; theta = 0.25; sigma_v = 0.1
h = 1
rrho(rho_old,v,y,mu,k,theta,sigma_v,h)
}
\seealso{
\code{\link[=rrho_proposal]{rrho_proposal()}}, \code{\link[=drho_proposal]{drho_proposal()}} for information about the
proposal distribution \eqn{P(\cdot)}.
}
