% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rsigma_v.R
\name{crsigma_v}
\alias{crsigma_v}
\alias{rsigma_v}
\title{Update \eqn{\sigma_v}}
\usage{
crsigma_v(sigma_v_old, prior_shape, prior_rate, v, y, mu, k, theta, rho, h)

rsigma_v(sigma_v_old, prior_shape, prior_rate, v, y, mu, k, theta, rho, h)
}
\arguments{
\item{sigma_v_old}{value of \eqn{\sigma_v} before updating, i.e.,
\eqn{\sigma_v^{(g)}}, in \eqn{g}th iteration.}

\item{prior_shape}{parameter shape of the prior of \eqn{\sigma_v}, (assumed)
gamma distribution.}

\item{prior_rate}{parameter rate of the prior of \eqn{\sigma_v}, (assumed)
gamma distribution.}

\item{v}{vector of volatility [\eqn{v_0, v_1, ..., v_N}].}

\item{y}{vector of returns [\eqn{y_1, ..., y_N}].}

\item{mu}{parameter \eqn{\mu}.}

\item{k}{parameter k.}

\item{theta}{parameter \eqn{\theta}.}

\item{rho}{parameter \eqn{\rho}.}

\item{h}{time unit.}
}
\value{
new value for \eqn{\sigma_v}, a scale value.
}
\description{
Update \eqn{\sigma_v}, using Independence Metropolis with a gamma
distribution as the
proposal distribution. Meanwhile, the prior distribution of \eqn{\sigma_v}
is also (assumed) a gamma distribution with \emph{prior_shape} and \emph{prior_rate}.
Note that these two prior parameters should be chosen carefully to avoid
\code{Inf*0} in the middle of the computation. The prior density shouldn't be
close to 0 for the effective range of \eqn{\sigma_v}.
\itemize{
\item \code{crsigma_v()} implemented in C++, through package Rcpp.
\item \code{rsigma_v()} implemented in R.
}
}
\section{Major steps}{

\enumerate{
\item First, compute the rate parameter of the proposal distribution
\eqn{P(\cdot)} (gamma distribution, the shape parameter defaults to 2). Since
\deqn{var(v_n-v_{n-1})\approx\sigma_v^2E[v_{n-1}]h,}
we have
\deqn{\sigma_v^2\approx \frac{S^2_{v_n-v_{n-1}}}{\bar{v}_{n-1}h}}
where \eqn{S^2_{v_n-v_{n-1}}}, \eqn{\bar{v}_{n-1}} are the
corresponding sample variance and mean, respectively.
If we want the proposal has a mode
equals to square root of above approximation, we can set the proposal
rate as
\deqn{\beta_{proposal} = 1/
      \sqrt{S^2_{v_n-v_{n-1}}/(\bar{v}_{n-1}h)}.}
Note that mode of gamma distribution equals \eqn{(\alpha -1)/\beta} for
shape \eqn{\alpha \ge 1} and rate \eqn{\beta}.
\strong{Warning}: \eqn{S_{v_n-v_{n-1}}^2=0} happens in extreme case!
\item Then, propose a new \eqn{\sigma_v^{(g+1)}} according to gamma
distribution, \code{rgamma(1,shape=2,rate)}, where rate = \eqn{\beta_{proposal}}.
The proposal density ratio
\deqn{\frac{P(\sigma_v^{(g)})}{P(\sigma_v^{(g+1)})}
   =min\left( \frac{dgamma(\sigma_v^{(g)}, shape=2, rate=1/\beta_{proposal})}
   {dgamma(\sigma_v^{(g+1)}, shape=2, rate=1/\beta_{proposal})}
   \right).}
\item Last, accept \eqn{\sigma_v^{(g+1)}} with probability
\deqn{\alpha(\sigma_v^{(g)},\sigma_v^{(g+1)})
=min\left(\frac{\pi(\sigma_v^{(g+1)})}{\pi(\sigma_v^{(g)})}
         \times \frac{P(\sigma_v^{(g)})}{P(\sigma_v^{(g+1)})}, 1\right).}
}
}

\section{Posterior density ratio}{


Employ following version Heston SV model
\deqn{y_n=\mu h - \frac{1}{2}v_{n-1}h + \sqrt{v_{n-1}h}\epsilon_n^y}
\deqn{v_n-v_{n-1}=k(\theta-v_{n-1})h + \sigma_v\sqrt{v_{n-1}h}
      (\rho\epsilon_n^y + \sqrt{1-\rho^2}\epsilon_n)}
The posterior of \eqn{\sigma_v}
\deqn{\begin{matrix}
  P(\sigma_v|v_{0:N},y_{1:N})
  &\propto& P(y_{1:N},v_{0:N})\cdot P_{prior}(\sigma_v)\\
  &\propto&\prod_{n=1}^{N}(P(v_n|v_{n-1},y_n)P(y_n|v_{n-1}))
   \cdot P_{prior}(\sigma_v)\\
  &\propto&\prod_{n=1}^{N}P(v_n|v_{n-1},y_n)\cdot P_{prior}(\sigma_v)\\
  &\propto&\prod_{n=1}^{N}P(x_n|v_{n-1},y_n)\cdot P_{prior}(\sigma_v)
\end{matrix}}
where \eqn{x_n\triangleq v_n-v_{n-1}-k(\theta-v_{n-1})h},
\eqn{P_{prior}(\sigma_v)}
denotes the prior density of \eqn{\sigma_v}. We have
\deqn{x_n|v_{n-1},y_n\sim \mathcal{N}(
     \sigma_v\rho\sqrt{v_{n-1}h}\epsilon_n^y,
     \sigma_v^2(1-\rho^2)v_{n-1}h)}
where
\deqn{\sqrt{v_{n-1}h}\epsilon_n^y=y_n-\mu h+\frac{1}{2}v_{n-1}h.}
\deqn{P(x_n|v_{n-1},y_n) = \frac{1}{\sigma_v\sqrt{2\pi(1-\rho^2)v_{n-1}h}}
      e^{-\frac{(x_n-\sigma_v\rho\sqrt{v_{n-1}h}\epsilon_n^y)^2}
      {2\sigma_v^2(1-\rho^2)v_{n-1}h}}.}

Then the posterior density ratio
\deqn{\frac{\pi(\sigma_v^{(g+1)})}
           {\pi(\sigma_v^{(g)})} =
      \frac{P(\sigma_v^{(g+1)}|v_{0:N},y_{1:N})}
           {P(\sigma_v^{(g)}|v_{0:N},y_{1:N})}.}
\deqn{log\left(\frac{\pi(\sigma_v^{(g+1)})}{\pi(\sigma_v^{(g)})}\right)
  = - N\cdot log\frac{\sigma_v^{(g+1)}}{\sigma_v^{(g)}}
  - \sum_{n=1}^N\left[
  \frac{(x_n-\sigma_v^{(g+1)}\rho\sqrt{v_{n-1}h}\epsilon_n^y)^2}
  {2\sigma_v^{(g+1)2}(1-\rho^2)v_{n-1}h}
  - \frac{(x_n-\sigma_v^{(g)}\rho\sqrt{v_{n-1}h}\epsilon_n^y)^2}
  {2\sigma_v^{(g)2}(1-\rho^2)v_{n-1}h}\right]
  + log\frac{P_{prior}(\sigma_v^{(g+1)})}{P_{prior}(\sigma_v^{(g)})}.}
}

\examples{
sigma_v_old = 0.01; prior_shape = 2; prior_rate = 1
v = c(0.25,0.2,0.1,0.15,0.25) # variance of diff(v) can't be 0!
y = rep(0.125,4)
mu = 0.125; k = 0.1; theta = 0.25; rho = 0
h = 1
rsigma_v(sigma_v_old,prior_shape,prior_rate,v,y,mu,k,theta,rho,h)
}
