% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rvN.R
\name{crvN}
\alias{crvN}
\alias{rvN}
\title{Update \eqn{v_N}}
\usage{
crvN(vN_old, vNm1, yN, mu, k, theta, sigma_v, rho, h)

rvN(vN_old, vNm1, yN, mu, k, theta, sigma_v, rho, h)
}
\arguments{
\item{vN_old}{value of \eqn{v_N} before updating, i.e.,
\eqn{v_N^{(g)}}, in the \eqn{g}th iteration.}

\item{vNm1}{volatility \eqn{v_{N-1}}.}

\item{yN}{return \eqn{y_N}.}

\item{mu}{parameter \eqn{\mu}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma_v}{parameter \eqn{\sigma_v}.}

\item{rho}{parameter \eqn{\rho}.}

\item{h}{time unit.}
}
\value{
new value of \eqn{v_N}, a scale value.
}
\description{
Generate new last volatility \eqn{v_N} according to its posterior,
using Independence Metropolis with a gamma density proposal \eqn{P(\cdot)},
i.e., \code{rgamma(1, shape=2, rate=1/vN_old)}, whose mode equals \emph{vN_old}.
\itemize{
\item \code{crvN()} implemented in C++, through package Rcpp.
\item \code{rvN()} implemented in R.
}
}
\section{Major steps}{

\enumerate{
\item First, propose a new \eqn{v_N^{(g+1)}} according to gamma distribution,
\code{rgamma(1, shape=2, rate=1/vN_old)}.
\item Then, compute the proposal density ratio
\deqn{\frac{P(v_N^{(g)})}{P(v_N^{(g+1)})}=
         \frac{dgamma(v_N^{(g)},shape=2,rate=1/v_N^{(g+1)})}
              {dgamma(v_N^{(g+1)},shape=2,rate=1/v_N^{(g)})}.}
\item Last, accept \eqn{v_N^{(g+1)}} with probability
\deqn{\alpha(v_N^{(g)},v_N^{(g+1)})
      =min\left(\frac{\pi(v_N^{(g+1)})}{\pi(v_N^{(g)})}
          \times \frac{P(v_N^{(g)})}{P(v_N^{(g+1)})},1\right).}
}
}

\section{Posterior density ratio}{


Employ following version Heston SV model
\deqn{y_n=\mu h -\frac{1}{2}v_{n-1}h + \sqrt{v_{n-1}h}\epsilon_n^y}
\deqn{v_n-v_{n-1}=k(\theta-v_{n-1})h + \sigma_v\sqrt{v_{n-1}h}(
     \rho\epsilon_n^y + \sqrt{1-\rho^2}\epsilon_n)}
The posterior
\deqn{\begin{matrix}
       P(v_N|v_{0:N-1},y_{1:N})
       &\propto& P(v_{0:N},y_{1:N})\\
       &\propto& P(v_N|v_{N-1},y_N)
      \end{matrix}}
Introduce notation
\deqn{\tilde{v}_N\triangleq v_{N-1}+k(\theta-v_{N-1})h+\sigma_v\rho
      \sqrt{v_{N-1}h}\epsilon_N^y,\quad
      \sqrt{v_{N-1}h}\epsilon_N^y = y_N-\mu h +\frac{1}{2}v_{N-1}h}
thus, we have
\deqn{v_{N}|v_{N-1},y_{N}
      \sim \mathcal{N}(\tilde{v}_{N}, \sigma_v^2(1-\rho^2)v_{N-1}h),\quad
      P(v_{N}|v_{N-1},y_{N})
      = \frac{1}{\sqrt{2\pi(1-\rho^2)v_{N-1}h}\sigma_v}
        e^{-\frac{(v_{N}-\tilde{v}_{N})^2}{2\sigma_v^2(1-\rho^2)v_{N-1}h}}.}
The posterior density ratio
\deqn{\begin{matrix}
  \frac{\pi(v_N^{(g+1)})}{\pi(v_N^{(g)})}
    &=& \frac{P(v_N^{(g+1)}|v_{N-1},y_N)}{P(v_N^{(g)}|v_{N-1},y_N)}\\
  log\left(\frac{\pi(v_N^{(g+1)})}{\pi(v_N^{(g)})}\right)
    &=& -\frac{(v_N^{(g+1)}-\tilde{v}_N)^2}{2\sigma_v^2(1-\rho^2)hv_{N-1}}
        +\frac{(v_N^{(g)}-\tilde{v}_N)^2}{2\sigma_v^2(1-\rho^2)hv_{N-1}}
\end{matrix}}
}

\examples{
v = rep(0.25,5); y = rep(0.125,4)
mu = 0.01; k = 0.1; theta = 0.25; sigma_v = 0.1; rho = 0; h = 1
N = length(y)
rvN(v[N+1],v,y,mu,k,theta,sigma_v,rho,h)
}
