% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_series}
\alias{Y_series}
\title{Simulated returns of Heston Stochastic Volatility Model data}
\format{
\subsection{\code{Y_series}}{

A vector with 1,000 elements:
\describe{
\item{Y[n]}{sample of \eqn{y_n}}
}
}
}
\source{
Simulated using Euler approximation
}
\usage{
Y_series
}
\description{
A vector of returns with length 1,000, simulated using Euler approximation
with true parameter setting:
\deqn{(\mu=0.125,k=0.1,\theta=0.25,\sigma_v=0.1,\rho=-0.7), h=1}
where \eqn{h} denotes time unit.
}
\keyword{datasets}
